python-scripts/requirements.txt
requests 
pygithub3 
PyGithub
gitpython
pathlib
